sorting and other functinalities are done

+++ src/pages/MyToys/MyToys.jsx
@@ -1,8 +1,10 @@
 import React, { useContext, useEffect, useState } from 'react';
 import useTitleChange from '../../TitleChange/TitleChange';
 import { AuthContext } from '../../providers/AuthProvider';
-import ToyRow from '../AllToys/ToyRow/ToyRow';
 import MyToyRow from './MyToyRow/MyToyRow';
+
+import Swal from 'sweetalert2';
+import { FaArrowDown, FaArrowUp } from 'react-icons/fa';
 
 const MyToys = () => {
   useTitleChange('Legowelt || My Toys');
@@ -11,29 +13,59 @@ const MyToys = () => {
 
   useEffect(() => {
     const fetchMyToys = async () => {
-      const response = await fetch(`http://localhost:5000/myToy/${user.email}`);
+      const response = await fetch(
+        `https://legowelt-server.vercel.app/myToy/${user?.email}`
+      );
       const data = await response.json();
       setMyToys(data);
     };
     fetchMyToys();
-  }, []);
+  }, [user]);
+
+  const sortHighPrice = () => {
+    fetch(`https://legowelt-server.vercel.app/sortHighPrice/${user?.email}`)
+      .then((res) => res.json())
+      .then((data) => {
+        setMyToys(data);
+      });
+  };
+
+  const sortLowPrice = () => {
+    fetch(`https://legowelt-server.vercel.app/sortLowPrice/${user?.email}`)
+      .then((res) => res.json())
+      .then((data) => {
+        setMyToys(data);
+      });
+  };
 
   // delete the toy
 
   const deleteToy = (id) => {
-    fetch(`http://localhost:5000/delete-toy/${id}`, {
-      method: 'DELETE',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        if (data.deletedCount > 0) {
-          alert('Toy deleted!');
-          setMyToys(myToys.filter((toy) => toy._id !== id));
-        }
-      });
+    Swal.fire({
+      title: 'Are you sure?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`https://legowelt-server.vercel.app/delete-toy/${id}`, {
+          method: 'DELETE',
+          headers: {
+            'Content-Type': 'application/json',
+          },
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            if (data.deletedCount > 0) {
+              setMyToys(myToys.filter((toy) => toy._id !== id));
+            }
+          });
+
+        Swal.fire('Deleted!', 'Your file has been deleted.', 'success');
+      }
+    });
   };
 
   return (
@@ -44,8 +76,12 @@ const MyToys = () => {
             Name
           </div>
 
-          <div className="flex flex-col text-[#0077b6] font-medium text-base md:text-xl">
+          <div className="flex gap-2 md:gap-8 text-[#0077b6] font-medium text-base md:text-xl items-center">
             Price
+            <div className="flex gap-3 md:gap-5">
+              <FaArrowUp onClick={sortHighPrice}></FaArrowUp>
+              <FaArrowDown onClick={sortLowPrice}></FaArrowDown>
+            </div>
           </div>
           <div className="flex flex-col text-[#0077b6] font-medium text-base md:text-xl">
             Update
